move connect from List to Main
